+++ src/bindings/yew.rs
@@ -2,8 +2,6 @@
 //! The user doesn't need to do anything but to put a style into the class of a
 //! yew component.
 
-extern crate yew;
-
 use crate::Style;
 use yew::html::Classes;