+++ src/style.rs
@@ -4,14 +4,15 @@ use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
 
 use once_cell::sync::Lazy;
-#[cfg(target_arch = "wasm32")]
-use wasm_bindgen::prelude::*;
-#[cfg(target_arch = "wasm32")]
-use web_sys::Element;
 
 use crate::ast::{Scope, ToCss};
 use crate::parser::Parser;
 use crate::utils::get_rand_str;
+#[cfg(target_arch = "wasm32")]
+use crate::utils::{doc_head, document};
+#[cfg(target_arch = "wasm32")]
+use crate::Error;
+use crate::Result;
 
 static STYLE_REGISTRY: Lazy<Arc<Mutex<StyleRegistry>>> = Lazy::new(|| Arc::new(Mutex::default()));
 
@@ -33,6 +34,70 @@ struct StyleContent {
     style_str: OnceCell<String>,
 }
 
+impl StyleContent {
+    fn get_class_name(&self) -> &str {
+        &self.class_name
+    }
+
+    fn get_style_str(&self) -> &str {
+        self.style_str.get_or_init(|| {
+            self.ast
+                .iter()
+                .map(|scope| scope.to_css(self.get_class_name()))
+                .fold(String::new(), |mut acc, css_part| {
+                    acc.push('\n');
+                    acc.push_str(&css_part);
+                    acc
+                })
+        })
+    }
+
+    /// Mounts the styles to the document
+    #[cfg(target_arch = "wasm32")]
+    fn mount(&self) -> Result<()> {
+        let document = document()?;
+        let head = doc_head()?;
+
+        let style_element = document
+            .create_element("style")
+            .map_err(|e| Error::Web(Some(e)))?;
+        style_element
+            .set_attribute("data-style", self.get_class_name())
+            .map_err(|e| Error::Web(Some(e)))?;
+        style_element.set_text_content(Some(self.get_style_str()));
+
+        head.append_child(&style_element)
+            .map_err(|e| Error::Web(Some(e)))?;
+        Ok(())
+    }
+
+    /// Unmounts the style from the DOM tree
+    /// Does nothing if it's not in the DOM tree
+    #[cfg(target_arch = "wasm32")]
+    fn unmount(&self) -> Result<()> {
+        let document = document()?;
+
+        if let Some(m) = document
+            .query_selector(&format!("style[data-style={}]", self.class_name))
+            .map_err(|e| Error::Web(Some(e)))?
+        {
+            if let Some(parent) = m.parent_element() {
+                parent.remove_child(&m).map_err(|e| Error::Web(Some(e)))?;
+            }
+        }
+
+        Ok(())
+    }
+}
+
+#[cfg(target_arch = "wasm32")]
+impl Drop for StyleContent {
+    /// Unmounts the style from the HTML head web-sys style
+    fn drop(&mut self) {
+        let _result = self.unmount();
+    }
+}
+
 #[derive(Debug, Clone)]
 pub struct Style {
     inner: Arc<StyleContent>,
@@ -40,20 +105,20 @@ pub struct Style {
 
 impl Style {
     /// Creates a new style
-    pub fn new<S: Into<Cow<'static, str>>>(css: S) -> Result<Self, String> {
+    pub fn new<S: Into<Cow<'static, str>>>(css: S) -> Result<Self> {
         Self::create("stylist", css)
     }
 
     /// Returns the class name for current style
     pub fn get_class_name(&self) -> &str {
-        &self.inner.class_name
+        self.inner.get_class_name()
     }
 
     /// Creates a new style with custom class prefix
     pub fn create<I1: AsRef<str>, I2: Into<Cow<'static, str>>>(
         class_prefix: I1,
         css: I2,
-    ) -> Result<Style, String> {
+    ) -> Result<Style> {
         let (class_prefix, css) = (class_prefix.as_ref(), css.into());
 
         let ast = Parser::parse(&*css)?;
@@ -66,7 +131,7 @@ impl Style {
         };
 
         #[cfg(target_arch = "wasm32")]
-        new_style.mount();
+        new_style.inner.mount()?;
 
         let style_registry_mutex = Arc::clone(&STYLE_REGISTRY);
         let mut style_registry = match style_registry_mutex.lock() {
@@ -82,62 +147,12 @@ impl Style {
 
     /// Get the parsed and generated style in `&str`.
     pub fn get_style_str(&self) -> &str {
-        self.inner.style_str.get_or_init(|| {
-            self.inner
-                .ast
-                .iter()
-                .map(|scope| scope.to_css(self.get_class_name()))
-                .fold(String::new(), |mut acc, css_part| {
-                    acc.push('\n');
-                    acc.push_str(&css_part);
-                    acc
-                })
-        })
+        self.inner.get_style_str()
     }
 }
 
 #[cfg(target_arch = "wasm32")]
-impl Style {
-    /// Mounts the styles to the document head web-sys style
-    fn mount(&self) {
-        if let Ok(node) = self.generate_element() {
-            let window = web_sys::window().expect("no global `window` exists");
-            let document = window.document().expect("should have a document on window");
-            let head = document.head().expect("should have a head in document");
-            head.append_child(&node).ok();
-        }
-    }
-
-    // Unmounts the style from the HTML head web-sys style
-    // fn unmount(&mut self) -> Self {
-    //    let window = web_sys::window().expect("no global `window` exists");
-    //    let document = window.document().expect("should have a document on window");
-    //
-    //   if let Some(m) = document
-    //       .query_selector(&format!("style[data-style={}]", self.class_name))
-    //       .ok()
-    //       .and_then(|m| m)
-    //   {
-    //       if let Some(parent) = m.parent_element() {
-    //           let _result = parent.remove_child(&m);
-    //       }
-    //   }
-    //   self.clone()
-    //}
-
-    /// Generates the `<style/>` tag web-sys style for inserting into the head of the
-    /// HTML document.
-    fn generate_element(&self) -> Result<Element, JsValue> {
-        let window = web_sys::window().expect("no global `window` exists");
-        let document = window.document().expect("should have a document on window");
-        let style_element = document.create_element("style").unwrap();
-        style_element
-            .set_attribute("data-style", self.get_class_name())
-            .ok();
-        style_element.set_text_content(Some(self.get_style_str()));
-        Ok(style_element)
-    }
-}
+impl Style {}
 
 impl ToString for Style {
     /// Just returns the classname